Apply order_required on view_that_asks_for_money
This view needs an order to work properly, otherwise it will cause an
AttributeError in the shop API when trying to get the order total via
get_order_total. Applying the order_required decorator fixes this by making
sure the client is redirected to the cart when there is no order on the
request.